Fix issues with migrations that are creating named parts
The `Container` & `Media` widgets are adding named parts, however the
migrations aren't quite set up right.

Also, by default the `FlowPart` isn't set to be reusable. The migration
will now make it reusable so it can be used as a named part. This also
improves the editing user interface as a title and description is
displayed.