Change widgets to use fields instead of custom parts
When first creating this module we were pretty new to Orchard Core.
After a few months and working on a few different projects we've come
to learn that we use fields more often than creating custom parts (in O1
we favoured using custom parts over fields).

This change is pretty much a complete rewrite of all the widgets to use
fields instead of custom parts. One downside to this is any projects
using this module already will not be able to upgrade, however going
forward it'll be easier to add new widgets and extend existing ones.
This approach is more in line with the work we're doing on our Orchard
Core projects.

Another noticeable change is instead of using migrations in code, the
content types are defined within a recipe (`widgets.recipe.json`).
Recipes are easy to construct (via deployment steps) than trying to
define content definitions through code within a migrations.

The `Section` content type is now using a `BagPart` instead of a
`FlowPart`. The `BagPart` gives us the ability to restrict which Items
can be addeed as a child of section, whereas the `FlowPart` will only
include all content types with stereotype of "Widget".

The view templates that control the output for the custom content parts
have obviously been removed as the content parts no longer exist. In
our projects we've been using Liquid templates over Razor. The new
templates control the output for the widget shape and are using Liquid.

The `HtmlAttributesPart` remains as this might be useful and can be
attached to a content type as and when the need arises. The new liquid
templates work without and with the `HtmlAttributesPart` attached.