No longer create content definitions as part of migration
If the content definitions for the widgets provided are required they
should either be run manually or be executed as part of another recipe.
When working on projects we often modify the common widget content
definitions, which can often require removing certain fields. To
prevent these being re-added when the module is enabled, the content
definition recipe no longer runs automatically.

Fixes #25.