Add "BackgroundPart"
Added a new part for storing background properties. Below are the
different fields that have been added with their intention.

Colour - Using our colour field for choosing a colour from a curated
palette or going rogue and using a custom colour
Pattern - Collection of patterns that point to images that should be
used as a background image
Invert Text - When using a dark background (or light if the theme is
dark by default), this flag indicates that the content should use
inverted colours to avoid bleding with background
- Fixed - When using a pattern, this can be used to give a parallax
effct