Add gallery widget
In a bid to sunset the separate gallery module, here's a simpler gallery as part of the widgets module that also resolves some longstanding issues (it's columnable and has alt text and appropriate width/height behaviour for flexible shaped images). There's an adjacent PR to the theme boilerplate for the required style changes.

Please note that once released, we should remove the existing gallery module from the siteboilerplate to remove conflicts