Fix Visual Level saving issue
When set to required, you can't save a Heading with "Match Level" as the option, because it is empty. Match Level is also the default so this will occur frequently until fixed. A temporary workaround is to pick the desired visual level on every heading but we'll want to get this sorted before major content population because it's annoying!