Rely on columns helper class
An adjacent PR adjusts the CSS in the ThemeBoilerplate so that Gallery doesn't need to have a re-implementation of column behaviour - it can use the helpers/_columns.scss helper classes instead. This change ensures the relevant helper classes are also output, and standardises the gallery with the way that other columnable widgets (like Content Feed and Tiles) work.