Ensure backwards compatability on templates that now support new link parts
Existing projects that use this module will use the various fields that
define link URLs and behaviour, these need to still utilised in the
templates to ensure the link functionality isn't broken. Each template
will check the existing fields first, then use the link parts if to
override variables if the parts are present and populated.